+++ pkg/webhook/admission/fake_validator_test.go
@@ -14,53 +14,53 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package admissiontest
+package admission
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
-// FakeValidator provides fake validating webhook functionality for testing
+// fakeValidator provides fake validating webhook functionality for testing
 // It implements the admission.Validator interface and
 // rejects all requests with the same configured error
 // or passes if ErrorToReturn is nil.
 // And it would always return configured warning messages WarningsToReturn.
-type FakeValidator struct {
-	// ErrorToReturn is the error for which the FakeValidator rejects all requests
+type fakeValidator struct {
+	// ErrorToReturn is the error for which the fakeValidator rejects all requests
 	ErrorToReturn error `json:"errorToReturn,omitempty"`
 	// GVKToReturn is the GroupVersionKind that the webhook operates on
 	GVKToReturn schema.GroupVersionKind
-	// WarningsToReturn is the warnings for FakeValidator returns to all requests
+	// WarningsToReturn is the warnings for fakeValidator returns to all requests
 	WarningsToReturn []string
 }
 
-func (v *FakeValidator) ValidateCreate() (warnings []string, err error) {
+func (v *fakeValidator) ValidateCreate() (warnings []string, err error) {
 	return v.WarningsToReturn, v.ErrorToReturn
 }
 
-func (v *FakeValidator) ValidateUpdate(old runtime.Object) (warnings []string, err error) {
+func (v *fakeValidator) ValidateUpdate(old runtime.Object) (warnings []string, err error) {
 	return v.WarningsToReturn, v.ErrorToReturn
 }
 
-func (v *FakeValidator) ValidateDelete() (warnings []string, err error) {
+func (v *fakeValidator) ValidateDelete() (warnings []string, err error) {
 	return v.WarningsToReturn, v.ErrorToReturn
 }
 
-func (v *FakeValidator) SetGroupVersionKind(gvk schema.GroupVersionKind) {
+func (v *fakeValidator) SetGroupVersionKind(gvk schema.GroupVersionKind) {
 	v.GVKToReturn = gvk
 }
 
-func (v *FakeValidator) GroupVersionKind() schema.GroupVersionKind {
+func (v *fakeValidator) GroupVersionKind() schema.GroupVersionKind {
 	return v.GVKToReturn
 }
 
-func (v *FakeValidator) GetObjectKind() schema.ObjectKind {
+func (v *fakeValidator) GetObjectKind() schema.ObjectKind {
 	return v
 }
 
-func (v *FakeValidator) DeepCopyObject() runtime.Object {
-	return &FakeValidator{
+func (v *fakeValidator) DeepCopyObject() runtime.Object {
+	return &fakeValidator{
 		ErrorToReturn:    v.ErrorToReturn,
 		GVKToReturn:      v.GVKToReturn,
 		WarningsToReturn: v.WarningsToReturn,
